int to bigint

+++ database/migrations/2019_08_12_092002_create_user_addresses_table.php
@@ -15,7 +15,7 @@ class CreateUserAddressesTable extends Migration
     {
         Schema::create('user_addresses', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->string('name',280)->nullable();            
             $table->smallInteger('area_id')->nullable();            
             $table->string('map_coordinates',480)->nullable();
